requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django>=4.0.2
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests>=2.20.0
lxml>=4.6.5
django-environ==0.8.1
Pillow>=9.0.0
PyYAML==6.0
drf-yasg>=1.20.0
django-filter>=2.4.0
django-cors-headers>=3.8.0
djangorestframework-simplejwt>=4.3.0
python-dateutil==2.8.2
djoser==2.1.0
